requirements.txt
openai==1.12.0
python-dotenv==1.0.1
click==8.1.7 